Flatten sync event handling and remove dead helpers
