+++ packages/code-link-cli/src/helpers/watcher.ts
@@ -8,8 +8,7 @@ import { isSupportedExtension, normalizePath, sanitizeFilePath } from "@code-lin
 import chokidar from "chokidar"
 import fs from "fs/promises"
 import path from "path"
-import { createScheduler } from "../scheduler.ts"
-import { TIMINGS } from "../timings.ts"
+import { createScheduler, TIMINGS } from "../scheduler.ts"
 import type { WatcherEvent } from "../types.ts"
 import { debug, warn } from "../utils/logging.ts"
 import { getRelativePath } from "../utils/node-paths.ts"
