+++ packages/code-link-cli/src/runtime.ts
@@ -2,9 +2,8 @@ import { randomUUID } from "node:crypto"
 import { normalizeCodeFilePathWithExtension, type PromptSession, type SyncPhase } from "@code-link/shared"
 import path from "path"
 import type { Installer } from "./helpers/installer.ts"
-import { createScheduler } from "./scheduler.ts"
+import { createScheduler, TIMINGS } from "./scheduler.ts"
 import { type FileOperationMemory, SyncMemory } from "./sync-memory.ts"
-import { TIMINGS } from "./timings.ts"
 import type { Conflict } from "./types.ts"
 import type { FileMetadataCache, FileSyncMetadata } from "./utils/file-metadata-cache.ts"
 import type { PersistedFileState } from "./utils/state-persistence.ts"
